+++ mediadecode/src/decoder.rs
@@ -175,7 +175,7 @@ pub trait SubtitleDecoder {
 mod tests {
   use super::*;
   use crate::Timebase;
-  use core::num::NonZeroU32;
+  use core::num::NonZeroI32;
 
   pub(crate) struct VLoop;
   impl VideoAdapter for VLoop {
@@ -225,7 +225,7 @@ mod tests {
       0
     }
     fn frame_rate(&self) -> Timebase {
-      Timebase::new(30, NonZeroU32::new(1).unwrap())
+      Timebase::new(30, NonZeroI32::new(1).unwrap())
     }
     fn duration(&self) -> Timestamp {
       Timestamp::new(0, self.frame_rate())